requirements.txt
finance-datareader>=0.9.96
pandas>=2.2.0
jinja2>=3.1.0
PyYAML>=6.0.0
pytest>=8.0.0
requests>=2.33.0
